+++ apps/docs/contentlayer.config.js
@@ -180,11 +180,7 @@ export default makeSource({
             [rehypePrettyCode, rehypeOptions],
             () => tree => {
                 visit(tree, node => {
-                    if (node?.type === "element" && node?.tagName === "div") {
-                        if (!("data-rehype-pretty-code-fragment" in node.properties)) {
-                            return;
-                        }
-
+                    if (node?.type === "element" && node?.tagName === "figure") {
                         const titleChild = node.children.find(child => {
                             return (
                                 child.properties &&